+++ sdk/storage/azblob/internal/exported/transfer_validation_option.go
@@ -13,25 +13,20 @@ import (
 	"io"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/streaming"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/internal/generated"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/internal/shared"
 )
 
-// OptionsSetter is a helper for setting the appropriate config field (NOT exported).
-type OptionsSetter interface {
-	SetCRC64([]byte)
-	SetMD5([]byte)
-}
-
 // TransferValidationType abstracts the various mechanisms used to verify a transfer.
 type TransferValidationType interface {
-	Apply(io.ReadSeekCloser, OptionsSetter) (io.ReadSeekCloser, error)
+	Apply(io.ReadSeekCloser, generated.TransactionalContentSetter) (io.ReadSeekCloser, error)
 	notPubliclyImplementable()
 }
 
 // TransferValidationTypeCRC64 is a TransferValidationType used to provide a precomputed CRC64.
 type TransferValidationTypeCRC64 uint64
 
-func (c TransferValidationTypeCRC64) Apply(rsc io.ReadSeekCloser, cfg OptionsSetter) (io.ReadSeekCloser, error) {
+func (c TransferValidationTypeCRC64) Apply(rsc io.ReadSeekCloser, cfg generated.TransactionalContentSetter) (io.ReadSeekCloser, error) {
 	buf := make([]byte, 8)
 	binary.LittleEndian.PutUint64(buf, uint64(c))
 	cfg.SetCRC64(buf)
@@ -42,7 +37,7 @@ func (TransferValidationTypeCRC64) notPubliclyImplementable() {}
 
 // TransferValidationTypeComputeCRC64 is a TransferValidationType that indicates a CRC64 should be computed during transfer.
 func TransferValidationTypeComputeCRC64() TransferValidationType {
-	return transferValidationTypeFn(func(rsc io.ReadSeekCloser, cfg OptionsSetter) (io.ReadSeekCloser, error) {
+	return transferValidationTypeFn(func(rsc io.ReadSeekCloser, cfg generated.TransactionalContentSetter) (io.ReadSeekCloser, error) {
 		buf, err := io.ReadAll(rsc)
 		if err != nil {
 			return nil, err
@@ -56,16 +51,16 @@ func TransferValidationTypeComputeCRC64() TransferValidationType {
 // TransferValidationTypeMD5 is a TransferValidationType used to provide a precomputed MD5.
 type TransferValidationTypeMD5 []byte
 
-func (c TransferValidationTypeMD5) Apply(rsc io.ReadSeekCloser, cfg OptionsSetter) (io.ReadSeekCloser, error) {
+func (c TransferValidationTypeMD5) Apply(rsc io.ReadSeekCloser, cfg generated.TransactionalContentSetter) (io.ReadSeekCloser, error) {
 	cfg.SetMD5(c)
 	return rsc, nil
 }
 
 func (TransferValidationTypeMD5) notPubliclyImplementable() {}
 
-type transferValidationTypeFn func(io.ReadSeekCloser, OptionsSetter) (io.ReadSeekCloser, error)
+type transferValidationTypeFn func(io.ReadSeekCloser, generated.TransactionalContentSetter) (io.ReadSeekCloser, error)
 
-func (t transferValidationTypeFn) Apply(rsc io.ReadSeekCloser, cfg OptionsSetter) (io.ReadSeekCloser, error) {
+func (t transferValidationTypeFn) Apply(rsc io.ReadSeekCloser, cfg generated.TransactionalContentSetter) (io.ReadSeekCloser, error) {
 	return t(rsc, cfg)
 }