+++ sdk/storage/azblob/CHANGELOG.md
@@ -17,6 +17,8 @@
 * CRC64 transactional hashes are now supplied with a `uint64` rather than a `[]byte` to conform with Golang's `hash/crc64` package
 * Field `XMSContentCRC64` has been renamed to `ContentCRC64`
 * Fields `TransactionalContentCRC64` and `TransactionalContentMD5` have been replaced by `TransactionalValidation`.
+* Fields `SourceContentCRC64` and `SourceContentMD5` have been replaced by `SourceContentValidation`.
+* Field `TransactionalContentMD5` has been removed from type `AppendBlockFromURLOptions`.
 
 ### Bugs Fixed